requirements.txt
# GUI and HTML dependencies
customtkinter
premailer
tkhtmlview
customtkinter
WeasyPrint
google-generativeai
premailer
jinja2>=3.1.2
tkhtmlview
pyinstaller
Pillow
markdown
openai

